+++ BambooTracker/gui/pattern_editor/pattern_editor_panel.hpp
@@ -120,6 +120,7 @@ signals:
 
 	void selected(bool isSelected);
 	void instrumentEntered(int num);
+	void volumeEntered(int volume);
 	void effectEntered(QString text);
 
 protected: